fix: ORM hardening — driver-aware identifier quoting, __get/__set attributes, count() int return, limit(offset,count) semantics, error-path finish/reset, create() immutability, find() scalar guard

+++ demo/application/Models/Admin/Group.php
@@ -21,7 +21,7 @@ class Group extends \Gene\Orm\Model
     function lists($start, $pagesize)
     {
         $where = ['group_pid' => 0];
-        $count = (int) static::query()->where($where)->count();
+        $count = static::query()->where($where)->count();
         $list = static::query()
             ->where($where)
             ->order('group_id asc')
